Build: make build error fix
This patch is for fixing build error like below.

sometimes the build error occur which
obj/startup.d cannot be found.
The cause is that the obj folder cannot be found by
compiling without performing mkdir first in
the multi-thread environment like -j16.
so the code was modified to execute mkdir first.

Signed-off-by: chungho.shin <chungho.shin@nexell.co.kr>
Change-Id: I04f2ecb868572e2671fdf771c6e660b44eee4118
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/bl1_s5p4418/+/488939
Tested-by: Nexell Jenkins <jenkins@nexell.co.kr>
Reviewed-by: allan park <allan.park@nexell.co.kr>